Renamed files and moved them around to make more sense and give better structure.
