+++ scripts/core/editor3/components/images/ImageBlock.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import {connect} from 'react-redux';
-import * as actions from '../../../actions';
+import * as actions from '../../actions';
 import {Entity} from 'draft-js';
 
 /**
