+++ scripts/core/editor3/components/images/ImageButton.jsx
@@ -1,20 +1,20 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import * as actions from '../../../actions';
+import * as actions from '../../actions';
 
 /**
  * @ngdoc React
  * @module superdesk.core.editor3
- * @name ImageControl
+ * @name ImageButton
  * @param {Function} insertImages Dispatches the insert images action.
  * @description This component renders the image button on the toolbar.
  */
-const ImageControlComponent = ({insertImages}) =>
+const ImageButtonComponent = ({insertImages}) =>
     <div className="Editor3-styleButton">
-        <span onClick={insertImages}>image</span>
+        <span onClick={insertImages}>img</span>
     </div>;
 
-ImageControlComponent.propTypes = {
+ImageButtonComponent.propTypes = {
     insertImages: React.PropTypes.func.isRequired
 };
 
@@ -22,6 +22,6 @@ const mapDispatchToProps = (dispatch) => ({
     insertImages: () => dispatch(actions.insertImages())
 });
 
-const ImageControl = connect(null, mapDispatchToProps)(ImageControlComponent);
+const ImageButton = connect(null, mapDispatchToProps)(ImageButtonComponent);
 
-export default ImageControl;
+export default ImageButton;
