+++ scripts/core/editor3/components/links/entityUtils.js
